Helpers: move third party dependencies to helper
I switched this function to a local, because Suzuki's Readrum has a
function with the same name, which pollutes the global scope. In those
instances of global scope pollution, issues can arise, such as lua
trying to run the function from the wrong place…
Switching to a local remediates the issue.